requirements/packages.txt
Flask==0.10.1
redis==2.9.1
PyYAML==3.10

